requirements.txt
setuptools~=68.2.0
chardet~=5.2.0
pillow~=10.4.0
tensorflow~=2.17.0
numpy~=1.26.4
keras~=3.6.0
matplotlib~=3.9.2
pandas~=2.2.3